Use SIZE_MAX consistently for invalid event handler indices
The event handler reference is a size_t, and the static initializer
in pmix_event.h already marks an invalid index with SIZE_MAX, but
the runtime constructor and all the registration error paths used
UINT_MAX instead - a different value on LP64 platforms. Standardize
on SIZE_MAX. Callers are expected to key off the returned status,
not the sentinel index, so this has no behavioral effect beyond
consistency.

+++ src/event/pmix_event_notification.c
@@ -1590,7 +1590,7 @@ pmix_status_t pmix_prep_event_chain(pmix_event_chain_t *chain, const pmix_info_t
 static void sevcon(pmix_event_hdlr_t *p)
 {
     p->name = NULL;
-    p->index = UINT_MAX;
+    p->index = SIZE_MAX;
     p->precedence = PMIX_EVENT_ORDER_NONE;
     p->oneshot = false;
     p->locator = NULL;
